+++ Wms.WebApp/Components/Pages/InventoryTransferPages/Details.razor.cs
@@ -35,7 +35,7 @@ public partial class Details
     private static string FormatDateTimeOffset(DateTimeOffset? value) =>
         value?.ToLocalTime().ToString("dd.MM.yyyy HH:mm") ?? "—";
 
-    private static string FormatQuantity(double value) => value.ToString("0.###");
+    private static string FormatQuantity(decimal value) => value.ToString("0.###");
 
     private string GetUserName(string? userId) => string.IsNullOrWhiteSpace(userId)
         ? "—"
