+++ Wms.WebApp/Components/Pages/InventoryTransferPages/Work.razor.cs
@@ -33,15 +33,15 @@ public partial class Work
     private StorageLocation? _pickSource;
     private List<InventoryTransferStorageLocationBalance> _pickSourceBalances = [];
     private Guid? _pickSkuId;
-    private double _pickQuantity;
+    private decimal _pickQuantity;
     private Guid? _putSkuId;
     private StorageLocation? _putDestination;
-    private double _putQuantity;
+    private decimal _putQuantity;
     private StorageLocation? _directSource;
     private List<InventoryTransferStorageLocationBalance> _directSourceBalances = [];
     private StorageLocation? _directDestination;
     private Guid? _directSkuId;
-    private double _directQuantity;
+    private decimal _directQuantity;
 
     private bool _isLoading = true;
     private bool _isSaving;
@@ -322,7 +322,7 @@ public partial class Work
     private static string FormatDateTimeOffset(DateTimeOffset? value) =>
         value?.ToLocalTime().ToString("dd.MM.yyyy HH:mm") ?? "—";
 
-    private static string FormatQuantity(double value) => value.ToString("0.###");
+    private static string FormatQuantity(decimal value) => value.ToString("0.###");
 
     private string GetUserName(string? userId) => string.IsNullOrWhiteSpace(userId)
         ? "—"
