+++ Wms/Application/Inventory/Transfers/InventoryTransferStorageLocationBalance.cs
@@ -6,6 +6,6 @@ namespace Wms.Application.Inventory.Transfers;
 public class InventoryTransferStorageLocationBalance
 {
     public StockKeepingUnit StockKeepingUnit { get; init; } = null!;
-    public double Quantity { get; init; }
+    public decimal Quantity { get; init; }
     public double? WeightKg => WeightCalculation.CalculateKg(Quantity, StockKeepingUnit);
 }
